ISSUE #366 modularise runCommand
