+++ tools/bouncer_worker/src/constants/errorCodes.js
@@ -1,0 +1,7 @@
+module.exports = {
+	ERRCODE_BOUNCER_CRASH: 12,
+	ERRCODE_PARAM_READ_FAIL: 13,
+	ERRCODE_BUNDLE_GEN_FAIL: 14,
+	ERRCODE_TIMEOUT: 29,
+	BOUNCER_SOFT_FAILS: [7,10,15] //failures that should go through to generate bundle
+}
